requirements.txt
black==19.10b0
isort==4.3.21
pytest==5.3.5
pytest-cov==2.8.1
Django==3.0.3
djangorestframework==3.11.0
